requirements.txt
fastapi 
uvicorn 
psycopg2 
asyncpg 
sqlalchemy
psycopg2-binary
python-dotenv
pytest 
httpx
pytest-asyncio
